+++ HC_Webapp/src/controllers/controller.cart.js
@@ -2,22 +2,21 @@ const { validationResult } = require('express-validator');
 const {
   writeCartJSON,
   indexCart,
-} = require('../model/products.model');
+} = require('../model/cart.model');
 
 const controllerCart = {
   mostrarCart: (req, res) => {
     return res.render('products/cart');
   }, 
     productsCart : (req, res) => {
-      
       let cartList = indexCart();
       let i = req.params.id;
       res.render('products/cart', {
         cartList: cartList,
         i: i,
-        
       });
-    }
+    },
+  
 } 
  
 module.exports = controllerCart;
