+++ HC_Webapp/src/model/products.model.js
@@ -7,11 +7,7 @@ const productModel = {
     let productJSON = readFileSync(productFile);
     return JSON.parse(productJSON);
   },
-  indexCart: function () {
-    let productFile = resolve(__dirname, '../data', 'cartList.json');
-    let cartJSON = readFileSync(productFile);
-    return JSON.parse(cartJSON);
-  },
+ 
   createProduct: function (data, imageName) {
     let products = productModel.indexProduct();
     return Object({
@@ -64,11 +60,7 @@ const productModel = {
     let update = JSON.stringify(data, null, 2);
     return writeFileSync(productFile, update);
   },
-  writeCartJSON: function (data) {
-    let productFile = resolve(__dirname, '../data', 'cartList.json');
-    let update = JSON.stringify(data, null, 2);
-    return writeFileSync(productFile, update);
-  },
+  
 };
 
 module.exports = productModel;
